move Option to its own file
